// frontend/src/shared/lib/api/model/paramFieldResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * prot-cellar
 * OpenAPI spec version: 0.1.0
 */
import type { ParamFieldResponseDefault } from './paramFieldResponseDefault';
import type { ParamFieldResponseHelp } from './paramFieldResponseHelp';

export interface ParamFieldResponse {
  key: string;
  label: string;
  type: string;
  required: boolean;
  default?: ParamFieldResponseDefault;
  options?: string[];
  help?: ParamFieldResponseHelp;
}